refactor: attitude 컴포넌트 디렉토리명 main으로 수정
