+++ components/main/AttitudeTitle.tsx
