+++ components/main/SplashRabbit.tsx
