+++ components/main/index.ts
